Patron Abstract Factory implementado para la creación de balas o proyectiles.

+++ core/src/com/mygdx/game/proyectil/ProyectilFactory.java
@@ -1,5 +1,5 @@
 package com.mygdx.game.proyectil;
 
 public interface ProyectilFactory {
-    public Proyectil crearProyectil(int x, int y, int vel, int rot, int dmg);
+    public Proyectil crearProyectil(float x, float y, int vel, int rot, int dmg);
 }
